requirements.txt
Django==2.0.13
django-appconf==1.0.2
django-bootstrap4==0.0.8
django-cities-light
django-tenants==2.1.0
django-select2==6.0.3
gunicorn==19.8.1
Pillow==5.2.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.4
reportlab==3.5.19
six==1.11.0
send==0.2.0
social-auth-app-django==3.1.0
whitenoise==3.3.1
django-widget-tweaks==1.4.3
requests-oauthlib==1.0.0
sqlparse==0.3.0
stripe==1.82.1
urllib3==1.24.2
braintree==3.53.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
django-allauth==0.36.0
idna==2.7